Key Press Done

+++ src/Paragraph.java
@@ -3,7 +3,7 @@ import java.util.List;
 public class Paragraph extends Song{
     private List<Syllable> syllables;
 
-    public Paragraph (String text, int speed) {
+    public Paragraph (String text, Float speed) {
         super(text, speed);
     }
